Threading
This was an experiment to see how threading would play out.

It works. But the overhead isn't worth the gain for this use-case, so it
ran about 4 times slower. There are likely optimisations that can be
done (Eg putting the reader in a separate thread and allocating more
memory to make it unrestrained), but I don't think it will be worth it.

Still, it was really cool to play around with.

I'm not going to invest time into making the code nice.